requirements.txt
# 核心依赖
PyQt5>=5.15.0
numpy>=1.19.0
Pillow>=8.0.0

# Habitat-sim相关依赖
# 注意：habitat-sim需要通过conda安装或从源码编译
# habitat-sim
# magnum (通常与habitat-sim一起安装)

# 可选：用于更好的字体支持
# matplotlib>=3.3.0 (如果需要更好的字体渲染)
